main: Move migratedumpsuccess to own file
